Add images to local slideshows.
Also, add randomized start times for videos.
Rename load video -> load local, because that's what it does.
